+++ lib/domain/repository/partner.dart
@@ -31,10 +31,10 @@ import 'paginated.dart';
 /// [MyUser] partner repository interface.
 abstract class AbstractPartnerRepository {
   /// Returns the balance [MyUser] has in their partner available wallet.
-  Rx<Balance> get available;
+  Rx<Balance?> get available;
 
   /// Returns the balance [MyUser] has in their partner hold wallet.
-  Rx<Balance> get hold;
+  Rx<Balance?> get hold;
 
   /// Returns the [Operation]s happening in [MyUser]'s partner wallet.
   Paginated<OperationId, Rx<Operation>> get operations;