+++ lib/ui/page/home/page/my_profile/view.dart
@@ -470,9 +470,11 @@ Widget _block(BuildContext context, MyProfileController c, int i) {
           FieldButton(
             onPressed: () async {
               // Don't show a confirmation when e-mail is set.
-              if (c.myUser.value?.emails.confirmed.isNotEmpty != true) {
-                await ConfirmLogoutView.show(router.context!);
+              if (c.myUser.value?.emails.confirmed.isNotEmpty == true) {
+                return c.logout();
               }
+
+              await ConfirmLogoutView.show(router.context!);
             },
             child: Text('btn_logout'.l10n),
           ),