+++ CHANGELOG.md
@@ -11,13 +11,23 @@ All user visible changes to this project will be documented in this file. This p
 
 [Diff](/../../compare/v0.4.1...v0.4.2) | [Milestone](/../../milestone/14)
 
+### Added
+
+- UI:
+    - Withdraw page:
+        - Dummy USDC and SWIFT withdrawal methods. ([#55])
+
 ### Changed
 
 - UI:
     - Wallet tab:
         - Searching and transactions order on transactions page. ([#54])
+        - Balance placeholder skeleton animation. ([#55])
+    - Monetization tab:
+        - Balance placeholder skeleton animation. ([#55])
 
 [#54]: /../../pull/54
+[#55]: /../../pull/55